some ambient loops tweak
