+++ Resources/Audio/_CP14/Ambience/Loops/attributions.yml
@@ -1,4 +1,4 @@
-- files: ["grasslands.ogg"]
+- files: ["winds.ogg"]
   license: "CC-BY-NC-4.0"
   copyright: 'by jgrzinich of Freesound.org.'
   source: "https://freesound.org/people/jgrzinich/sounds/329805/"
